refactor(oauth): extract scope data into a dependency-free module
`OAUTH_PROVIDERS` imports ~120 React icon components and reads `env` at
module load, so nothing outside Next can read the scope sets it declares.
That blocks the docs generator from publishing what each integration
requests, and it buries a security-relevant surface inside a file that
also owns token refresh and error redaction.

Move the 61 per-service scope arrays out of `oauth.ts`, and the scope
label tables plus `getScopeDescription` out of `utils.ts`, into a new
`lib/oauth/scopes.ts` with no imports at all. `oauth.ts` composes from it
(`scopes: [...OAUTH_SCOPES.sharepoint]`), and Slack's approval-gated
extras stay appended there because whether they are requested depends on
an environment flag.

Behavior-preserving: every one of the 61 scope lists was verified
identical before and after the move. Consumers of the moved exports are
repointed at the new module rather than re-exported through `utils.ts`,
which the import rules forbid.

The new test asserts the two files cover exactly the same services and
that every scope in the table is actually requested at authorization
time, so the split cannot silently drift.
