+++ apps/sim/lib/oauth/oauth.ts
@@ -86,6 +86,7 @@ import {
   SALESFORCE_PROVIDER_ID_LABELS,
 } from '@/lib/oauth/salesforce'
 import { REDDIT_USER_AGENT } from '@/tools/reddit/constants'
+import { OAUTH_SCOPES } from './scopes'
 import type { OAuthProviderConfig } from './types'
 
 const logger = createLogger('OAuth')
@@ -114,7 +115,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'claude-platform-service-account',
         icon: ClaudeIcon,
         baseProviderIcon: ClaudeIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES['claude-platform']],
         authType: 'service_account',
       },
     },
@@ -130,13 +131,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-email',
         icon: GmailIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/gmail.send',
-          'https://www.googleapis.com/auth/gmail.modify',
-          'https://www.googleapis.com/auth/gmail.labels',
-        ],
+        scopes: [...OAUTH_SCOPES.gmail],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-drive': {
@@ -145,12 +140,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-drive',
         icon: GoogleDriveIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/drive.file',
-          'https://www.googleapis.com/auth/drive',
-        ],
+        scopes: [...OAUTH_SCOPES['google-drive']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-docs': {
@@ -159,12 +149,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-docs',
         icon: GoogleDocsIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/drive.file',
-          'https://www.googleapis.com/auth/drive',
-        ],
+        scopes: [...OAUTH_SCOPES['google-docs']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-sheets': {
@@ -173,12 +158,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-sheets',
         icon: GoogleSheetsIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/drive.file',
-          'https://www.googleapis.com/auth/drive',
-        ],
+        scopes: [...OAUTH_SCOPES['google-sheets']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-forms': {
@@ -187,13 +167,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-forms',
         icon: GoogleFormsIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/drive',
-          'https://www.googleapis.com/auth/forms.body',
-          'https://www.googleapis.com/auth/forms.responses.readonly',
-        ],
+        scopes: [...OAUTH_SCOPES['google-forms']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-calendar': {
@@ -202,11 +176,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-calendar',
         icon: GoogleCalendarIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/calendar',
-        ],
+        scopes: [...OAUTH_SCOPES['google-calendar']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-contacts': {
@@ -215,11 +185,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-contacts',
         icon: GoogleContactsIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/contacts',
-        ],
+        scopes: [...OAUTH_SCOPES['google-contacts']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-ads': {
@@ -228,11 +194,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-ads',
         icon: GoogleAdsIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/adwords',
-        ],
+        scopes: [...OAUTH_SCOPES['google-ads']],
       },
       'google-bigquery': {
         name: 'Google BigQuery',
@@ -240,11 +202,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-bigquery',
         icon: GoogleBigQueryIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/bigquery',
-        ],
+        scopes: [...OAUTH_SCOPES['google-bigquery']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-tasks': {
@@ -253,11 +211,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-tasks',
         icon: GoogleTasksIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/tasks',
-        ],
+        scopes: [...OAUTH_SCOPES['google-tasks']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-vault': {
@@ -266,17 +220,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-vault',
         icon: GoogleVaultIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/ediscovery',
-          // Least-privilege scope for read-only consumers. The knowledge base
-          // connector only lists matters, holds, and saved queries, all of which
-          // accept ediscovery.readonly; the block's export tools still need the
-          // read-write scope above.
-          'https://www.googleapis.com/auth/ediscovery.readonly',
-          'https://www.googleapis.com/auth/devstorage.read_only',
-        ],
+        scopes: [...OAUTH_SCOPES['google-vault']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-groups': {
@@ -285,12 +229,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-groups',
         icon: GoogleGroupsIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/admin.directory.group',
-          'https://www.googleapis.com/auth/admin.directory.group.member',
-        ],
+        scopes: [...OAUTH_SCOPES['google-groups']],
         serviceAccountProviderId: 'google-service-account',
       },
       /**
@@ -307,12 +246,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-chat',
         icon: GoogleChatIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/chat.spaces.readonly',
-          'https://www.googleapis.com/auth/chat.messages.readonly',
-        ],
+        scopes: [...OAUTH_SCOPES['google-chat']],
       },
       'google-meet': {
         name: 'Google Meet',
@@ -320,12 +254,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-meet',
         icon: GoogleMeetIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/meetings.space.created',
-          'https://www.googleapis.com/auth/meetings.space.readonly',
-        ],
+        scopes: [...OAUTH_SCOPES['google-meet']],
         serviceAccountProviderId: 'google-service-account',
       },
       'google-service-account': {
@@ -334,7 +263,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'google-service-account',
         icon: GoogleIcon,
         baseProviderIcon: GoogleIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES['google-service-account']],
         authType: 'service_account',
       },
       'vertex-ai': {
@@ -343,11 +272,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'vertex-ai',
         icon: VertexIcon,
         baseProviderIcon: VertexIcon,
-        scopes: [
-          'https://www.googleapis.com/auth/userinfo.email',
-          'https://www.googleapis.com/auth/userinfo.profile',
-          'https://www.googleapis.com/auth/cloud-platform',
-        ],
+        scopes: [...OAUTH_SCOPES['vertex-ai']],
       },
     },
     defaultService: 'gmail',
@@ -362,24 +287,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'microsoft-ad',
         icon: AzureIcon,
         baseProviderIcon: MicrosoftIcon,
-        scopes: [
-          'openid',
-          'profile',
-          'email',
-          'User.ReadWrite.All',
-          'Group.ReadWrite.All',
-          'GroupMember.ReadWrite.All',
-          'LicenseAssignment.Read.All',
-          'LicenseAssignment.ReadWrite.All',
-          'UserAuthenticationMethod.ReadWrite.All',
-          'AuditLog.Read.All',
-          'Application.Read.All',
-          'AppRoleAssignment.ReadWrite.All',
-          'RoleManagement.ReadWrite.Directory',
-          'Device.Read.All',
-          'Policy.Read.All',
-          'offline_access',
-        ],
+        scopes: [...OAUTH_SCOPES['microsoft-ad']],
       },
       'microsoft-dataverse': {
         name: 'Microsoft Dataverse',
@@ -387,13 +295,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'microsoft-dataverse',
         icon: MicrosoftDataverseIcon,
         baseProviderIcon: MicrosoftIcon,
-        scopes: [
-          'openid',
-          'profile',
-          'email',
-          'https://dynamics.microsoft.com/user_impersonation',
-          'offline_access',
-        ],
+        scopes: [...OAUTH_SCOPES['microsoft-dataverse']],
       },
       'microsoft-excel': {
         name: 'Microsoft Excel',
@@ -401,7 +303,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'microsoft-excel',
         icon: MicrosoftExcelIcon,
         baseProviderIcon: MicrosoftIcon,
-        scopes: ['openid', 'profile', 'email', 'Files.Read', 'Files.ReadWrite', 'offline_access'],
+        scopes: [...OAUTH_SCOPES['microsoft-excel']],
       },
       'microsoft-planner': {
         name: 'Microsoft Planner',
@@ -409,15 +311,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'microsoft-planner',
         icon: MicrosoftPlannerIcon,
         baseProviderIcon: MicrosoftIcon,
-        scopes: [
-          'openid',
-          'profile',
-          'email',
-          'Group.ReadWrite.All',
-          'Group.Read.All',
-          'Tasks.ReadWrite',
-          'offline_access',
-        ],
+        scopes: [...OAUTH_SCOPES['microsoft-planner']],
       },
       'microsoft-teams': {
         name: 'Microsoft Teams',
@@ -425,28 +319,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'microsoft-teams',
         icon: MicrosoftTeamsIcon,
         baseProviderIcon: MicrosoftIcon,
-        scopes: [
-          'openid',
-          'profile',
-          'email',
-          'User.Read',
-          'Chat.Read',
-          'Chat.ReadWrite',
-          'Chat.ReadBasic',
-          'ChatMessage.Send',
-          'Channel.ReadBasic.All',
-          'ChannelMessage.Send',
-          'ChannelMessage.Read.All',
-          'ChannelMessage.ReadWrite',
-          'ChannelMember.Read.All',
-          'Group.Read.All',
-          'Group.ReadWrite.All',
-          'Team.ReadBasic.All',
-          'TeamMember.Read.All',
-          'offline_access',
-          'Files.Read',
-          'Sites.Read.All',
-        ],
+        scopes: [...OAUTH_SCOPES['microsoft-teams']],
       },
       'microsoft-word': {
         name: 'Microsoft Word',
@@ -454,30 +327,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'microsoft-word',
         icon: MicrosoftWordIcon,
         baseProviderIcon: MicrosoftIcon,
-        /**
-         * Word documents are ordinary drive items, so the integration reads and
-         * writes them through the Files permissions rather than a Word-specific
-         * scope — Microsoft Graph exposes no Word API of its own.
-         *
-         * The `.All` variants are what make the SharePoint drive the block
-         * exposes actually work: `Files.ReadWrite` alone covers only the signed-in
-         * user's own OneDrive, so a document library would be rejected for
-         * insufficient privileges. Both are user-consentable, so this does not
-         * push the integration behind admin consent, and neither grants access to
-         * anything the signed-in account could not already open.
-         *
-         * @see https://learn.microsoft.com/en-us/graph/permissions-reference
-         */
-        scopes: [
-          'openid',
-          'profile',
-          'email',
-          'Files.Read',
-          'Files.ReadWrite',
-          'Files.Read.All',
-          'Files.ReadWrite.All',
-          'offline_access',
-        ],
+        scopes: [...OAUTH_SCOPES['microsoft-word']],
       },
       outlook: {
         name: 'Outlook',
@@ -485,37 +335,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'outlook',
         icon: OutlookIcon,
         baseProviderIcon: MicrosoftIcon,
-        /**
-         * `Calendars.ReadWrite` backs the Outlook calendar operations. Graph documents it
-         * as the sole accepted permission for creating and updating events and for
-         * accept / tentativelyAccept / decline ("Higher: Not available"), and it is
-         * supported for both work/school and personal Microsoft accounts.
-         *
-         * Do NOT add `Calendars.ReadWrite.Shared` here. This provider is shared by work
-         * and personal Outlook accounts, and the `.Shared` calendar scopes are not
-         * confirmed supported for personal Microsoft accounts — requesting one risks
-         * failing consent for personal users, which would take mail access down with it.
-         * That is the same reasoning that kept `findMeetingTimes` out of this integration.
-         * The consequence is that calendar operations target calendars the account owns;
-         * picking a calendar shared by another user may return 403 from Graph.
-         *
-         * Microsoft only grants newly-added scopes on a fresh authorization, so users who
-         * connected Outlook before `Calendars.ReadWrite` existed must reconnect
-         * (re-consent) before the calendar operations will work.
-         *
-         * @see https://learn.microsoft.com/en-us/graph/permissions-reference
-         */
-        scopes: [
-          'openid',
-          'profile',
-          'email',
-          'Mail.ReadWrite',
-          'Mail.ReadBasic',
-          'Mail.Read',
-          'Mail.Send',
-          'Calendars.ReadWrite',
-          'offline_access',
-        ],
+        scopes: [...OAUTH_SCOPES.outlook],
       },
       onedrive: {
         name: 'OneDrive',
@@ -523,7 +343,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'onedrive',
         icon: MicrosoftOneDriveIcon,
         baseProviderIcon: MicrosoftIcon,
-        scopes: ['openid', 'profile', 'email', 'Files.Read', 'Files.ReadWrite', 'offline_access'],
+        scopes: [...OAUTH_SCOPES.onedrive],
       },
       sharepoint: {
         name: 'SharePoint',
@@ -531,15 +351,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'sharepoint',
         icon: MicrosoftSharepointIcon,
         baseProviderIcon: MicrosoftIcon,
-        scopes: [
-          'openid',
-          'profile',
-          'email',
-          'Sites.Read.All',
-          'Sites.ReadWrite.All',
-          'Sites.Manage.All',
-          'offline_access',
-        ],
+        scopes: [...OAUTH_SCOPES.sharepoint],
       },
     },
     defaultService: 'outlook',
@@ -554,23 +366,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'x',
         icon: xIcon,
         baseProviderIcon: xIcon,
-        scopes: [
-          'tweet.read',
-          'tweet.write',
-          'tweet.moderate.write',
-          'users.read',
-          'follows.read',
-          'follows.write',
-          'bookmark.read',
-          'bookmark.write',
-          'like.read',
-          'like.write',
-          'block.read',
-          'block.write',
-          'mute.read',
-          'mute.write',
-          'offline.access',
-        ],
+        scopes: [...OAUTH_SCOPES.x],
       },
     },
     defaultService: 'x',
@@ -585,13 +381,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'tiktok',
         icon: TikTokIcon,
         baseProviderIcon: TikTokIcon,
-        scopes: [
-          'user.info.basic',
-          'user.info.profile',
-          'user.info.stats',
-          'video.upload',
-          'video.list',
-        ],
+        scopes: [...OAUTH_SCOPES.tiktok],
       },
     },
     defaultService: 'tiktok',
@@ -607,7 +397,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'atlassian-service-account',
         icon: AtlassianIcon,
         baseProviderIcon: AtlassianIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES['atlassian-service-account']],
         authType: 'service_account',
       },
     },
@@ -624,44 +414,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         icon: ConfluenceIcon,
         baseProviderIcon: ConfluenceIcon,
         serviceAccountProviderId: 'atlassian-service-account',
-        scopes: [
-          'read:confluence-content.all',
-          'read:confluence-space.summary',
-          'read:space:confluence',
-          'write:confluence-content',
-          'write:confluence-space',
-          'write:confluence-file',
-          'read:page:confluence',
-          'write:page:confluence',
-          'read:comment:confluence',
-          'write:comment:confluence',
-          'delete:comment:confluence',
-          'delete:attachment:confluence',
-          'delete:page:confluence',
-          'read:label:confluence',
-          'write:label:confluence',
-          'read:attachment:confluence',
-          'write:attachment:confluence',
-          'search:confluence',
-          'read:me',
-          'offline_access',
-          'read:hierarchical-content:confluence',
-          'read:content.metadata:confluence',
-          'read:user:confluence',
-          'read:confluence-user',
-          'read:task:confluence',
-          'write:task:confluence',
-          'write:space:confluence',
-          'delete:space:confluence',
-          'read:blogpost:confluence',
-          'write:blogpost:confluence',
-          'delete:blogpost:confluence',
-          'read:content.property:confluence',
-          'write:content.property:confluence',
-          'read:space.property:confluence',
-          'write:space.property:confluence',
-          'read:space.permission:confluence',
-        ],
+        scopes: [...OAUTH_SCOPES.confluence],
       },
     },
     defaultService: 'confluence',
@@ -677,52 +430,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         icon: JiraIcon,
         baseProviderIcon: JiraIcon,
         serviceAccountProviderId: 'atlassian-service-account',
-        scopes: [
-          'read:jira-user',
-          'read:jira-work',
-          'write:jira-work',
-          'read:me',
-          'offline_access',
-          'read:issue.vote:jira',
-          'read:user:jira',
-          'delete:issue:jira',
-          'delete:comment:jira',
-          'delete:attachment:jira',
-          'delete:issue-worklog:jira',
-          'delete:issue-link:jira',
-          // Jira Service Management scopes. The classic scopes are required: Atlassian
-          // enforces an endpoint's granular scope set as all-of, and several JSM request
-          // endpoints include scopes outside this list in their granular sets.
-          'read:servicedesk-request',
-          'write:servicedesk-request',
-          'manage:servicedesk-customer',
-          'read:servicedesk:jira-service-management',
-          'read:requesttype:jira-service-management',
-          'read:request:jira-service-management',
-          'write:request:jira-service-management',
-          'read:request.comment:jira-service-management',
-          'write:request.comment:jira-service-management',
-          'read:servicedesk.customer:jira-service-management',
-          'write:servicedesk.customer:jira-service-management',
-          'read:organization:jira-service-management',
-          'write:organization:jira-service-management',
-          'read:servicedesk.organization:jira-service-management',
-          'write:servicedesk.organization:jira-service-management',
-          'read:queue:jira-service-management',
-          'read:request.sla:jira-service-management',
-          'read:request.status:jira-service-management',
-          'write:request.status:jira-service-management',
-          'read:request.participant:jira-service-management',
-          'write:request.participant:jira-service-management',
-          'read:request.approval:jira-service-management',
-          'write:request.approval:jira-service-management',
-          'read:cmdb-object:jira',
-          'write:cmdb-object:jira',
-          'delete:cmdb-object:jira',
-          'read:cmdb-schema:jira',
-          'read:cmdb-type:jira',
-          'read:cmdb-attribute:jira',
-        ],
+        scopes: [...OAUTH_SCOPES.jira],
       },
     },
     defaultService: 'jira',
@@ -738,13 +446,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'airtable-service-account',
         icon: AirtableIcon,
         baseProviderIcon: AirtableIcon,
-        scopes: [
-          'data.records:read',
-          'data.records:write',
-          'schema.bases:read',
-          'user.email:read',
-          'webhook:manage',
-        ],
+        scopes: [...OAUTH_SCOPES.airtable],
       },
     },
     defaultService: 'airtable',
@@ -759,16 +461,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'bitbucket',
         icon: BitbucketIcon,
         baseProviderIcon: BitbucketIcon,
-        scopes: [
-          'account',
-          'repository',
-          'repository:write',
-          'pullrequest',
-          'pullrequest:write',
-          'pipeline',
-          'pipeline:write',
-          'webhook',
-        ],
+        scopes: [...OAUTH_SCOPES.bitbucket],
       },
     },
     defaultService: 'bitbucket',
@@ -784,7 +477,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'notion-service-account',
         icon: NotionIcon,
         baseProviderIcon: NotionIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES.notion],
       },
     },
     defaultService: 'notion',
@@ -800,7 +493,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'clickup-service-account',
         icon: ClickUpIcon,
         baseProviderIcon: ClickUpIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES.clickup],
       },
     },
     defaultService: 'clickup',
@@ -816,7 +509,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'linear-service-account',
         icon: LinearIcon,
         baseProviderIcon: LinearIcon,
-        scopes: ['read', 'write'],
+        scopes: [...OAUTH_SCOPES.linear],
       },
     },
     defaultService: 'linear',
@@ -881,15 +574,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'monday-service-account',
         icon: MondayIcon,
         baseProviderIcon: MondayIcon,
-        scopes: [
-          'boards:read',
-          'boards:write',
-          'updates:read',
-          'updates:write',
-          'webhooks:read',
-          'webhooks:write',
-          'me:read',
-        ],
+        scopes: [...OAUTH_SCOPES.monday],
       },
     },
     defaultService: 'monday',
@@ -904,7 +589,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'box',
         icon: BoxCompanyIcon,
         baseProviderIcon: BoxCompanyIcon,
-        scopes: ['root_readwrite', 'sign_requests.readwrite'],
+        scopes: [...OAUTH_SCOPES.box],
         serviceAccountProviderId: 'box-service-account',
       },
     },
@@ -920,15 +605,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'dropbox',
         icon: DropboxIcon,
         baseProviderIcon: DropboxIcon,
-        scopes: [
-          'account_info.read',
-          'files.metadata.read',
-          'files.metadata.write',
-          'files.content.read',
-          'files.content.write',
-          'sharing.read',
-          'sharing.write',
-        ],
+        scopes: [...OAUTH_SCOPES.dropbox],
       },
     },
     defaultService: 'dropbox',
@@ -944,14 +621,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'shopify-service-account',
         icon: ShopifyIcon,
         baseProviderIcon: ShopifyIcon,
-        scopes: [
-          'write_products',
-          'write_orders',
-          'write_customers',
-          'write_inventory',
-          'read_locations',
-          'write_merchant_managed_fulfillment_orders',
-        ],
+        scopes: [...OAUTH_SCOPES.shopify],
       },
     },
     defaultService: 'shopify',
@@ -967,28 +637,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'slack-custom-bot',
         icon: SlackIcon,
         baseProviderIcon: SlackIcon,
-        scopes: [
-          'channels:read',
-          'channels:history',
-          'channels:manage',
-          'groups:read',
-          'groups:history',
-          'groups:write',
-          'chat:write',
-          'chat:write.public',
-          ...SLACK_APPROVAL_GATED_SCOPES,
-          'im:write',
-          'im:read',
-          'users:read',
-          // TODO: Add 'users:read.email' once Slack app review is approved
-          'files:write',
-          'files:read',
-          'canvases:read',
-          'canvases:write',
-          'reactions:write',
-          'reactions:read',
-          // TODO: Add 'pins:read' once Slack app review is approved
-        ],
+        scopes: [...OAUTH_SCOPES.slack, ...SLACK_APPROVAL_GATED_SCOPES],
       },
     },
     defaultService: 'slack',
@@ -1004,7 +653,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'snowflake-service-account',
         icon: SnowflakeIcon,
         baseProviderIcon: SnowflakeIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES.snowflake],
         authType: 'service_account',
       },
     },
@@ -1022,7 +671,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'netsuite-service-account',
         icon: NetSuiteIcon,
         baseProviderIcon: NetSuiteIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES.netsuite],
         authType: 'service_account',
       },
     },
@@ -1038,24 +687,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'reddit',
         icon: RedditIcon,
         baseProviderIcon: RedditIcon,
-        scopes: [
-          'identity',
-          'read',
-          'submit',
-          'vote',
-          'save',
-          'edit',
-          'subscribe',
-          'history',
-          'privatemessages',
-          'account',
-          'mysubreddits',
-          'flair',
-          'report',
-          'modposts',
-          'modflair',
-          'modmail',
-        ],
+        scopes: [...OAUTH_SCOPES.reddit],
       },
     },
     defaultService: 'reddit',
@@ -1071,7 +703,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'wealthbox-service-account',
         icon: WealthboxIcon,
         baseProviderIcon: WealthboxIcon,
-        scopes: ['login', 'data'],
+        scopes: [...OAUTH_SCOPES.wealthbox],
       },
     },
     defaultService: 'wealthbox',
@@ -1087,7 +719,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'webflow-service-account',
         icon: WebflowIcon,
         baseProviderIcon: WebflowIcon,
-        scopes: ['cms:read', 'cms:write', 'sites:read', 'sites:write', 'forms:read'],
+        scopes: [...OAUTH_SCOPES.webflow],
       },
     },
     defaultService: 'webflow',
@@ -1103,7 +735,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'trello-service-account',
         icon: TrelloIcon,
         baseProviderIcon: TrelloIcon,
-        scopes: ['read', 'write'],
+        scopes: [...OAUTH_SCOPES.trello],
       },
     },
     defaultService: 'trello',
@@ -1119,7 +751,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'asana-service-account',
         icon: AsanaIcon,
         baseProviderIcon: AsanaIcon,
-        scopes: ['default'],
+        scopes: [...OAUTH_SCOPES.asana],
       },
     },
     defaultService: 'asana',
@@ -1135,17 +767,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'attio-service-account',
         icon: AttioIcon,
         baseProviderIcon: AttioIcon,
-        scopes: [
-          'record_permission:read-write',
-          'object_configuration:read-write',
-          'list_configuration:read-write',
-          'list_entry:read-write',
-          'note:read-write',
-          'task:read-write',
-          'comment:read-write',
-          'user_management:read',
-          'webhook:read-write',
-        ],
+        scopes: [...OAUTH_SCOPES.attio],
       },
     },
     defaultService: 'attio',
@@ -1161,7 +783,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'calcom-service-account',
         icon: CalComIcon,
         baseProviderIcon: CalComIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES.calcom],
       },
     },
     defaultService: 'calcom',
@@ -1176,7 +798,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'docusign',
         icon: DocuSignIcon,
         baseProviderIcon: DocuSignIcon,
-        scopes: ['signature', 'extended'],
+        scopes: [...OAUTH_SCOPES.docusign],
       },
     },
     defaultService: 'docusign',
@@ -1192,15 +814,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'pipedrive-service-account',
         icon: PipedriveIcon,
         baseProviderIcon: PipedriveIcon,
-        scopes: [
-          'base',
-          'deals:full',
-          'contacts:full',
-          'leads:full',
-          'activities:full',
-          'mail:full',
-          'projects:full',
-        ],
+        scopes: [...OAUTH_SCOPES.pipedrive],
       },
     },
     defaultService: 'pipedrive',
@@ -1216,28 +830,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'hubspot-service-account',
         icon: HubspotIcon,
         baseProviderIcon: HubspotIcon,
-        scopes: [
-          'crm.objects.contacts.read',
-          'crm.objects.contacts.write',
-          'crm.objects.companies.read',
-          'crm.objects.companies.write',
-          'crm.objects.deals.read',
-          'crm.objects.deals.write',
-          'crm.objects.owners.read',
-          'crm.objects.users.read',
-          'crm.objects.marketing_events.read',
-          'crm.objects.line_items.read',
-          'crm.objects.line_items.write',
-          'crm.objects.quotes.read',
-          'crm.objects.appointments.read',
-          'crm.objects.appointments.write',
-          'crm.objects.carts.read',
-          'sales-email-read',
-          'crm.lists.read',
-          'crm.lists.write',
-          'tickets',
-          'oauth',
-        ],
+        scopes: [...OAUTH_SCOPES.hubspot],
       },
     },
     defaultService: 'hubspot',
@@ -1253,7 +846,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'harmonic-service-account',
         icon: HarmonicIcon,
         baseProviderIcon: HarmonicIcon,
-        scopes: [],
+        scopes: [...OAUTH_SCOPES.harmonic],
         authType: 'service_account',
       },
     },
@@ -1269,7 +862,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'linkedin',
         icon: LinkedInIcon,
         baseProviderIcon: LinkedInIcon,
-        scopes: ['profile', 'openid', 'email', 'w_member_social'],
+        scopes: [...OAUTH_SCOPES.linkedin],
       },
     },
     defaultService: 'linkedin',
@@ -1284,13 +877,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'instagram',
         icon: InstagramIcon,
         baseProviderIcon: InstagramIcon,
-        scopes: [
-          'instagram_business_basic',
-          'instagram_business_content_publish',
-          'instagram_business_manage_comments',
-          'instagram_business_manage_messages',
-          'instagram_business_manage_insights',
-        ],
+        scopes: [...OAUTH_SCOPES.instagram],
       },
     },
     defaultService: 'instagram',
@@ -1309,7 +896,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         serviceAccountProviderId: 'salesforce-service-account',
         icon: SalesforceIcon,
         baseProviderIcon: SalesforceIcon,
-        scopes: ['api', 'refresh_token', 'openid'],
+        scopes: [...OAUTH_SCOPES.salesforce],
       },
     },
     defaultService: 'salesforce',
@@ -1339,31 +926,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         // Desk.search.READ, Desk.webhooks.READ and Desk.webhooks.UPDATE were
         // requested but unused - no tool searches, and the provider never lists
         // or edits a subscription.
-        scopes: [
-          // READ + UPDATE rather than tickets.ALL: no tool creates or deletes a
-          // ticket, and ALL additionally grants ticket DELETE. Threads, comments
-          // and attachments live under the tickets module and are covered by
-          // these two. NOTE: Zoho publishes no scope line for the attachment
-          // content sub-path - verify attachment download against a live account
-          // before merge and widen here if it returns SCOPE_MISMATCH.
-          'Desk.tickets.READ',
-          'Desk.tickets.UPDATE',
-          'Desk.contacts.READ',
-          // READ only: the knowledge base connector syncs Help Center articles
-          // via GET /articles and GET /articles/{id}; nothing authors one.
-          'Desk.articles.READ',
-          // GET /organizations documents `Desk.organization.READ , Desk.basic.READ`.
-          // Sibling endpoints spell the same construction "requires X and Y"
-          // (dependencyMappings, roles), so the comma is AND, not OR.
-          'Desk.organization.READ',
-          // READ only: the agent picker for `assigneeId` lists agents, and no
-          // tool creates, edits or deletes one.
-          'Desk.agents.READ',
-          'Desk.basic.READ',
-          'Desk.webhooks.CREATE',
-          'Desk.webhooks.DELETE',
-          'aaaserver.profile.READ',
-        ],
+        scopes: [...OAUTH_SCOPES['zoho-desk']],
       },
     },
     defaultService: 'zoho-desk',
@@ -1378,19 +941,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'zoom',
         icon: ZoomIcon,
         baseProviderIcon: ZoomIcon,
-        scopes: [
-          'user:read:user',
-          'meeting:write:meeting',
-          'meeting:read:meeting',
-          'meeting:read:list_meetings',
-          'meeting:update:meeting',
-          'meeting:delete:meeting',
-          'meeting:read:invitation',
-          'meeting:read:list_past_participants',
-          'cloud_recording:read:list_user_recordings',
-          'cloud_recording:read:list_recording_files',
-          'cloud_recording:delete:recording_file',
-        ],
+        scopes: [...OAUTH_SCOPES.zoom],
         serviceAccountProviderId: 'zoom-service-account',
       },
     },
@@ -1406,7 +957,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'wordpress',
         icon: WordpressIcon,
         baseProviderIcon: WordpressIcon,
-        scopes: ['global'],
+        scopes: [...OAUTH_SCOPES.wordpress],
       },
     },
     defaultService: 'wordpress',
@@ -1421,25 +972,7 @@ export const OAUTH_PROVIDERS: Record<string, OAuthProviderConfig> = {
         providerId: 'spotify',
         icon: SpotifyIcon,
         baseProviderIcon: SpotifyIcon,
-        scopes: [
-          'user-read-private',
-          'user-read-email',
-          'user-library-read',
-          'user-library-modify',
-          'playlist-read-private',
-          'playlist-read-collaborative',
-          'playlist-modify-public',
-          'playlist-modify-private',
-          'user-read-playback-state',
-          'user-modify-playback-state',
-          'user-read-currently-playing',
-          'user-read-recently-played',
-          'user-top-read',
-          'user-follow-read',
-          'user-follow-modify',
-          'user-read-playback-position',
-          'ugc-image-upload',
-        ],
+        scopes: [...OAUTH_SCOPES.spotify],
       },
     },
     defaultService: 'spotify',
