+++ apps/sim/lib/oauth/utils.test.ts
@@ -8,7 +8,6 @@ import {
   getCanonicalScopesForProvider,
   getMissingRequiredScopes,
   getProviderIdFromServiceId,
-  getScopeDescription,
   getScopesForService,
   getServiceByProviderAndId,
   getServiceConfigByProviderId,
@@ -382,43 +381,6 @@ describe('getCanonicalScopesForProvider', () => {
 
     expect(Array.isArray(scopes)).toBe(true)
     expect(scopes.length).toBe(0)
-  })
-})
-
-describe('getScopeDescription', () => {
-  it.concurrent('uses provider-specific labels for Bitbucket scope names', () => {
-    expect(getScopeDescription('account', 'bitbucket')).toBe(
-      'View your Bitbucket account and workspace memberships'
-    )
-    expect(getScopeDescription('pipeline:write', 'bitbucket')).toBe('Run and stop pipelines')
-    expect(getScopeDescription('webhook', 'bitbucket')).toBe('Manage repository webhooks')
-  })
-
-  it.concurrent('preserves the existing Reddit meaning of the account scope', () => {
-    expect(getScopeDescription('account', 'reddit')).toBe('Update account preferences and settings')
-    expect(getScopeDescription('account')).toBe('Update account preferences and settings')
-  })
-
-  /**
-   * The consent screen is where a user decides what to grant, so a write scope
-   * has to read as one. `w_member_social` previously said 'Access LinkedIn
-   * profile', describing a posting grant as a profile read.
-   *
-   * The wording tracks LinkedIn's own: "Post, comment, and like posts on behalf
-   * of an authenticated member." It names all three verbs even though Sim only
-   * posts -- the label describes the grant the token carries, not Sim's current
-   * use of it, and LinkedIn's scopes cannot be sub-selected.
-   */
-  it.concurrent('describes w_member_social as the write grant it is', () => {
-    const description = getScopeDescription('w_member_social', 'linkedin')
-
-    expect(description).toBe('Post, comment, and like posts on your behalf')
-    expect(description).not.toMatch(/access .*profile/i)
-  })
-
-  it.concurrent('leaves the read-only LinkedIn scopes read-only', () => {
-    expect(getScopeDescription('profile', 'linkedin')).toBe('Access profile information')
-    expect(getScopeDescription('email', 'linkedin')).toBe('Access email address')
   })
 })
